# Code Quality: Improve string comparer performance