New comparator is slightly faster than [DllImport(“api-ms-win-core-string-l1-1-0.dll”)].CompareStringEx. And supports Arm. Previously, a string comparer was used for Arm without proper support for natural comparison.

| Method                                 | Mean     | Error     | StdDev    | Allocated |
|--------------------------------------- |---------:|----------:|----------:|----------:|
| Test_api-ms-win-core-string-l1-1-0.dll | 1.185 ms | 0.0095 ms | 0.0084 ms |      89 B |
| Test_NaturalComparer                   | 1.072 ms | 0.0073 ms | 0.0065 ms |      89 B |