@0x5bfa 2024-12-11
FYI, this isn't yet published and documented but C# runtime just superseded `CompareOptions.NumericOrdering` 2 weeks ago.