@0x5bfa 2024-12-11
FYI, C# runtime just superseded this with `CompareOptions.NumericOrdering` 2 weeks ago (not documented yet).