@0x5bfa 2024-12-12
C# runtime implemented SortOptions.NumericOrdering just two weeks ago. We will use it once we upgrade to NET 9.x, instead.